requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.3
pydantic~=2.11.5
SQLAlchemy~=2.0.41
passlib~=1.7.4
requests~=2.32.3
python-jose~=3.5.0